Updating OvershareKit and InAppSettingsKit.

+++ clients/ios/Other Sources/InAppSettingsKit/Models/IASKSettingsStoreUserDefaults.m
@@ -49,7 +49,7 @@
     [self.defaults setDouble:value forKey:key];
 }
 
-- (void)setInteger:(int)value forKey:(NSString*)key {
+- (void)setInteger:(NSInteger)value forKey:(NSString*)key {
     [self.defaults setInteger:value forKey:key];
 }
